try and make setup checks a bit more clear
